+++ plugins/system/sociallogin/layout/akeeba/sociallogin/button.php
@@ -6,6 +6,9 @@
  */
 
 // Protect from unauthorized access
+use Akeeba\SocialLogin\Library\Helper\Integrations;
+use Akeeba\SocialLogin\Library\Helper\Joomla;
+
 defined('_JEXEC') or die();
 
 /**
@@ -25,26 +28,59 @@ defined('_JEXEC') or die();
  * @var   string       $label       Text content of the button
  * @var   string       $icon_class  An icon class for the span inside the button
  * @var   string       $img         An <img> (or other) tag to use inside the button when $icon_class is empty
+ * @var   bool         $relocate    Should I try to move the social login button next to the regular login button?
+ * @var   string[]     $selectors   A list of CSS selectors I will use to find the regular login button in the module.
  */
 
+// BEGIN - MANDATORY CODE
+
 // Extract the data. Do not remove until the unset() line.
-extract(array_merge(array(
+extract(array_merge([
 	'slug'       => '',
 	'link'       => '',
 	'tooltip'    => '',
 	'label'      => '',
 	'icon_class' => '',
 	'img'        => '',
-), $displayData));
+	'relocate'   => false,
+	'selectors'  => [
+		"#form-login-submit > button",
+		"button[type=submit]",
+		"[type=submit]",
+		"[id*=\"submit\"]",
+	],
+], $displayData));
+
+$randomId = 'akeeba-sociallogin-' . Joomla::generateRandom(12) . '-' . Joomla::generateRandom(8);
+
+$jsSelectors = implode(", ", array_map(function ($selector) {
+	return '"' . addslashes($selector) . '"';
+}, $selectors));
+
+if ($relocate)
+{
+	Integrations::includeButtonRelocationJS();
+
+	$js = <<< JS
+
+window.jQuery(document).ready(function(){
+	akeeba_sociallogin_move_button(document.getElementById('{$randomId}'), [{$jsSelectors}]); 
+});
+
+JS;
+	Joomla::getApplication()->getDocument()->addScriptDeclaration($js);
+}
+// END - MANDATORY CODE
 
 // Start writing your template override code below this line
 ?>
-<a class="btn btn-default akeeba-sociallogin-button akeeba-sociallogin-button-<?php echo $slug?> hasTooltip"
-   href="<?php echo $link?>"  title="<?php echo $tooltip ?>">
-    <?php if (!empty($icon_class)): ?>
-    <span class="<?php echo $icon_class ?>"></span>
-    <?php else: ?>
-    <?php echo $img ?>
-    <?php endif; ?>
-    <?php echo $label ?>
+<a class="btn btn-default akeeba-sociallogin-button akeeba-sociallogin-button-<?= $slug ?> hasTooltip"
+   id="<?= $randomId ?>"
+   href="<?= $link ?>" title="<?= $tooltip ?>">
+	<?php if (!empty($icon_class)): ?>
+		<span class="<?= $icon_class ?>"></span>
+	<?php else: ?>
+		<?= $img ?>
+	<?php endif; ?>
+	<?= $label ?>
 </a>